owasp_api_demo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
bcrypt==4.1.2
httpx==0.25.2
pydantic[email]==2.5.1